create a new hook called useMovieDetails
