+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import ReactPaginate from 'react-paginate';
 import { Typography } from '@mui/material';
 import { MovieDetailsContext } from './context/MovieDetailsContext';
@@ -6,11 +6,8 @@ import MovieSearchBar from './components/MovieSearchBar';
 import MovieSearchResults from './components/MovieSearchResults/MovieSearchResults';
 import MovieDetails from './components/MovieDetails/MovieDetails';
 import useStyles from './App.styles';
-import type { MovieDetailsType } from './components/MovieDetails/MovieDetails.type';
-import { filterMovieDetailsData } from './components/MovieDetails/filterMovieDetailsKeys';
 import { useSearchResults } from './hooks/useSearchResults';
-
-const API_KEY = process.env.REACT_APP_API_KEY;
+import { useMovieDetails } from './hooks/useMovieDetails';
 
 const App = () => {
   const classes = useStyles();
@@ -25,11 +22,8 @@ const App = () => {
   const [searchQuery, setSearchQuery] = useState('');
 
   // MOVIE DETAILS
-  const [showMovieDetails, setShowMovieDetails] = useState(false);
-  const [shouldFetchMovieDetails, setShouldFetchMovieDetails] = useState(false);
   const [movieDetailsId, setMovieDetailsId] = useState('');
-  const [movieDetails, setMovieDetails] = useState({} as MovieDetailsType);
-  const [detailsDataIsLoading, setDetailsDataIsLoading] = useState(false);
+  const [showMovieDetails, setShowMovieDetails] = useState(false);
 
   // PAGINATION
   const [currentPage, setCurrentPage] = useState(1);
@@ -40,21 +34,11 @@ const App = () => {
     pageCount
   } = useSearchResults({searchQuery, currentPage, shouldFetchSearchResults, setShowNoResultsMessage, setSearchError});
 
-  const fetchApiMovieDetails = async (imdbID: string) => {
-    setShouldFetchMovieDetails(false);
-    setDetailsDataIsLoading(true);
-    const URL = `http://www.omdbapi.com/?apikey=${API_KEY}&i=${imdbID}&type=movie`;
-    try {
-      const response = await fetch(URL);
-      const movieDetails = await response.json();
 
-      const cleanMovieDetails = filterMovieDetailsData(movieDetails);
-      setMovieDetails(cleanMovieDetails);
-      setDetailsDataIsLoading(false);
-    } catch (e) {
-      console.log({ error: e });
-    }
-  };
+  const {
+    movieDetails,
+    detailsDataIsLoading
+  } = useMovieDetails(movieDetailsId)
 
   const onInputChange = (newValue: string) => {
     setInputValue(newValue);
@@ -74,7 +58,6 @@ const App = () => {
 
   const onUpdateMovieDetailsId = (id: string) => {
     setMovieDetailsId(id);
-    setShouldFetchMovieDetails(true);
   };
 
   const onPageChange = ({ selected }: { selected: number }) => {
@@ -82,11 +65,6 @@ const App = () => {
     setShouldFetchSearchResults(true);
   };
 
-  useEffect(() => {
-    if (shouldFetchMovieDetails) {
-      fetchApiMovieDetails(movieDetailsId);
-    }
-  }, [shouldFetchMovieDetails, movieDetailsId]);
 
   return (
     <div className={classes.root}>
