+++ src/hooks/types/UseSearchResults.type.ts
